application started without problems !!!

+++ src/main/java/com/student/StudentManagement/services/FiliereService.java
@@ -4,16 +4,18 @@ import com.student.StudentManagement.dto.RequestFiliereDto;
 import com.student.StudentManagement.model.Filiere;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface FiliereService {
     RequestFiliereDto createFiliere(RequestFiliereDto requestFiliereDto);
 
     List<Filiere> getAllFilieres();
 
-    Filiere getFiliereById(int id);
+    Filiere getFiliereById(Long id);
+
 
     void updateFiliere(Filiere filier);
 
-    void deleteFiliere(int id);
+    void deleteFiliere(Long id);
 }
